requirements.txt
langchain==1.0.4
langchain_community==0.4.1
langchain_core==1.0.3
langchain_google_genai==3.0.1
langchain_huggingface==1.0.1
langdetect==1.0.9
python-dotenv==1.2.1
streamlit==1.50.0
torch==2.8.0
youtube_transcript_api==1.2.3